requirements.txt
fastapi
uvicorn[standard]
httpx
openpyxl
python-dateutil
pydantic-settings
loguru
